+++ src/models/msg/msg_software_upgrade_proposal.ts
