+++ src/models/msg/msg_verify_invariant.ts
