+++ src/models/msg/msg_vote.ts
